refactor: 로또 생성 기능은 LottoTickets에서 LottoTicket로 이관
